chore: move activePlugins file
