+++ src/context/PluginProvider/PluginProvider.tsx
@@ -2,10 +2,10 @@ import { ChainId } from '@shapeshiftoss/caip'
 import { ChainAdapter } from '@shapeshiftoss/chain-adapters'
 import { KnownChainIds } from '@shapeshiftoss/types'
 import { PluginManager } from 'plugins'
+import { activePlugins } from 'plugins/activePlugins'
 import React, { createContext, useContext, useEffect, useMemo, useRef, useState } from 'react'
 import { useSelector } from 'react-redux'
 import { Route } from 'Routes/helpers'
-import { activePlugins } from 'context/PluginProvider/activePlugins'
 import { logger } from 'lib/logger'
 import { partitionCompareWith } from 'lib/utils'
 import { selectFeatureFlags } from 'state/slices/preferencesSlice/selectors'
